refactor: deprecated component definition.
