--- front/app/src/routes/_layout.component.tsx
@@ -1,8 +1,0 @@
-import { Outlet } from "@tanstack/react-router";
-import { Layout } from "@/components/layout";
-
-export const component = () => (
-  <Layout>
-    <Outlet />
-  </Layout>
-);
