+++ front/app/src/routes/_layout.tsx
@@ -1,0 +1,10 @@
+import {createFileRoute, Outlet} from "@tanstack/react-router";
+import { Layout } from "@/components/layout";
+
+export const Route = createFileRoute("/_layout")({
+    component: () => (
+        <Layout>
+            <Outlet />
+        </Layout>
+    )
+})
