--- front/app/src/routes/_layout/about/index.component.tsx
@@ -1,5 +1,0 @@
-import { Link } from "@tanstack/react-router";
-
-export function component() {
-  return <Link to="/">home</Link>;
-}
