+++ front/app/src/routes/_layout/about/index.tsx
@@ -1,0 +1,7 @@
+import {createFileRoute, Link} from "@tanstack/react-router";
+
+export const Route = createFileRoute("/_layout/about/")({
+  component: () => {
+    return <Link to="/">home</Link>;
+  }
+})
